requirements.txt
asgiref==3.10.0
azure-core==1.36.0
azure-storage-blob
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django>=5.0
django-environ==0.12.0
django-storages
django-taggit
gunicorn==23.0.0
idna==3.11
isodate==0.7.2
markdown
packaging==25.0
Pillow
psycopg[binary]
pycparser==2.23
python-dotenv
requests==2.32.5
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
whitenoise==6.11.0
